requirements.txt
# Backtrader投资分析系统依赖包

# 核心交易和回测框架
backtrader>=1.9.78.123

# 数据处理和分析
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# 数据库连接
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0

# HTTP请求和API调用
requests>=2.28.0
urllib3>=1.26.0

# 环境变量管理
python-dotenv>=0.19.0

# 图表和可视化
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0

# 数据获取
yfinance>=0.1.87
fredapi>=0.4.3
quandl>=3.7.0

# 技术分析
TA-Lib>=0.4.25
pandas-ta>=0.3.14b

# 机器学习（可选）
scikit-learn>=1.1.0

# 日期时间处理
python-dateutil>=2.8.0
pytz>=2022.1

# 配置文件处理
PyYAML>=6.0
configparser>=5.3.0

# 日志和调试
loguru>=0.6.0

# 数据验证

# AI和NLP依赖（媒体观点聚合功能）
transformers>=4.21.0
torch>=1.12.0
jieba>=0.42.1
feedparser>=6.0.10
pydantic>=1.10.0

# 异步处理
aiohttp>=3.8.0
aiofiles>=0.8.0

# 文本处理和报告生成
markdown>=3.4.0
jinja2>=3.1.0

# 数据序列化
orjson>=3.8.0

# 进度条
tqdm>=4.64.0

# 缓存
redis>=4.3.0

# 测试框架
pytest>=7.1.0
pytest-asyncio>=0.19.0

# 代码质量
flake8>=5.0.0
black>=22.6.0

# 类型检查
mypy>=0.971